Rename host files to RazorLight
